% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{add_motifs}
\alias{add_motifs}
\title{Motif to a sequence injection}
\usage{
add_motifs(motifs, sequence)
}
\arguments{
\item{motifs}{list of motifs to be injected}

\item{sequence}{vector of alphabet elements}
}
\value{
list(sequence, motifs, masks)
}
\description{
this function injects motifs to a sequence
}
\examples{
# simple injection
add_motifs(list(c(1, "_", 1), c(1, 1)), c(2, 2, 3, 4))
# little bit more interesting
alph <- as.character(1L:4)
motifs <- generate_motifs(alph, 2, 2, n = 4, d = 6)
example_sequence <- sample(alph, size = 10, replace = TRUE)
add_motifs(motifs, example_sequence)
}
