% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{contains_motif}
\alias{contains_motif}
\title{Simple motifs check}
\usage{
contains_motif(sequence, motifs)
}
\arguments{
\item{sequence}{a vector of characters}

\item{motifs}{list of motifs}
}
\value{
randomly generated sequences
}
\description{
This function check if at least one of provided motifs is contained in the
provided sequence.
}
\examples{
alph <- 1:4
motifs <- generate_motifs(alph, 3, 3, 3, 2)
sequence <- sample(alph,  100, replace = TRUE)
contains_motif(sequence, motifs)

}
