% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{count_seq_kmers}
\alias{count_seq_kmers}
\title{wrapper for seqR counters}
\usage{
count_seq_kmers(sequences, alphabet, n = 4, d = 6)
}
\arguments{
\item{sequences}{input data for k-mer counting}

\item{alphabet}{elements used to generate a motif}

\item{n}{maximum number of alphabet elements}

\item{d}{number of possible gaps}
}
\description{
wrapper for seqR counters
}
