% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_chisq}
\alias{filter_chisq}
\title{Chi-squared independence test}
\usage{
filter_chisq(target, kmers, significance_level = 0.05)
}
\arguments{
\item{target}{a numeric response variable}

\item{kmers}{a matrix of kmers with named columns or an object obtained via
\code{\link{generate_kmer_data}} function.}

\item{significance_level}{a number from 0-1 interval denoting significance
level for testing.}
}
\value{
a character vector of names of selected kmers
}
\description{
This function is a wrapper over chi-squared independence test for k-mer data.
}
\details{
This function uses \code{\link[stats]{chisq.test}}
}
\examples{
n_seq <- 20
sequence_length <- 20
alph <- letters[1:4]
motifs <- generate_motifs(alph, 4, 4, 4, 6)
kmers <- generate_kmer_data(n_seq, sequence_length, alph,
                            motifs, n_injections = 4)
target <- get_target_additive(kmers)
filter_chisq(target, kmers)

}
