% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_praznik}
\alias{filter_praznik}
\title{Entropy-Based Feature Selection Algorithms}
\usage{
filter_praznik(target, kmers, method, thresh)
}
\arguments{
\item{target}{a numeric response variable}

\item{kmers}{a matrix of kmers with named columns or an object obtained via
\code{\link{generate_kmer_data}} function.}

\item{method}{a character name of a filter type. One of "MIM", "MRMR", "JMI",
"JMIM", "DISR", "NJMIM", "CMIM", "CMI". See details for more information.}

\item{thresh}{a threshold for corresponfing score.}
}
\value{
a character vector of names of selected kmers
}
\description{
This function is a wrapper over Entropy-Based Feature Selection Algorithms
for k-mer data.
}
\details{
This function uses
- \code{\link[praznik]{MIM}}
- \code{\link[praznik]{MRMR}}
- \code{\link[praznik]{JMI}}
- \code{\link[praznik]{JMIM}}
- \code{\link[praznik]{DISR}}
- \code{\link[praznik]{NJMIM}}
- \code{\link[praznik]{CMIM}}
- \code{\link[praznik]{CMI}}
}
\examples{
n_seq <- 20
sequence_length <- 20
alph <- letters[1:4]
motifs <- generate_motifs(alph, 4, 4, 4, 6)
kmers <- generate_kmer_data(n_seq, sequence_length, alph,
                            motifs, n_injections = 4)
target <- get_target_additive(kmers)
filter_praznik(target, kmers, "infogain", 0.01)

}
