% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{generate_kmer_data}
\alias{generate_kmer_data}
\title{function counts n-grams in given sequences}
\usage{
generate_kmer_data(
  n_seq,
  sequence_length,
  alphabet,
  motifs,
  n_injections,
  fraction = 0.5,
  seqProbs = NULL,
  n = 4,
  d = 6
)
}
\arguments{
\item{n_seq}{number of sequences to be generated}

\item{sequence_length}{sequence length}

\item{alphabet}{elements used to build sequence}

\item{motifs}{list of motifs}

\item{n_injections}{number of motifs injected to each positive sequence}

\item{fraction}{fraction of positive sequences}

\item{seqProbs}{alphabet probabilites for sequences}

\item{n}{maximum number of alphabet elements in n-gram}

\item{d}{maximum number of gaps in n-gram}
}
\value{
generated sequences
}
\description{
function counts n-grams in given sequences
}
\examples{
n_seq <- 20
sequence_length <- 1200
alph <- letters[1:4]
motifs <- generate_motifs(alph, 2, 3, 4, 6)
results <- generate_kmer_data(n_seq, sequence_length, alph, motifs, 1)
}
