% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{generate_motifs}
\alias{generate_motifs}
\title{Function generates list of motifs}
\usage{
generate_motifs(alphabet, n_motifs, n, d, motifProbs = NULL, validate = TRUE)
}
\arguments{
\item{alphabet}{elements used to generate a motif}

\item{n_motifs}{number of motifs to generate}

\item{n}{maximum number of alphabet elements}

\item{d}{number of possible gaps}

\item{motifProbs}{alphabet elements' probabilites}

\item{validate}{if true, returns a set of motifs that can be injected to a sequence of length 10}
}
\value{
list of generated motifs
}
\description{
generate multiple motifs from alphabet
}
\examples{
generate_motifs(1:4, 5, n = 6, d = 6)
generate_motifs(1:4, 5, n = 6, d = 2, motifProbs = c(0.7, 0.1, 0.1, 0.1))
}
