% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{generate_negative_sequence}
\alias{generate_negative_sequence}
\title{Sampling from alphabet}
\usage{
generate_negative_sequence(
  sequence_length,
  alphabet,
  motifs,
  seqProbs = NULL,
  attempts = 100
)
}
\arguments{
\item{sequence_length}{sequence length}

\item{alphabet}{elements used to build sequence}

\item{motifs}{list of motifs}

\item{seqProbs}{alphabet probabilities for sequences}
}
\value{
randomly generated sequences
}
\description{
This function generates sequence of elements from alphabet with replacement
}
\examples{
set.seed(2)
alph <- 1:4
motifs <- generate_motifs(alph, 3, 3, 3, 2)
generate_negative_sequence(5, alph, motifs)
generate_negative_sequence(10, c("a", "b", "c"), motifs)
generate_negative_sequence(10, c("a", "b", "c"), c(0.6, 0.2, 0.2), motifs)

}
