% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbaasp_data.R
\name{get_dbaasp_data}
\alias{get_dbaasp_data}
\title{This function pulls the data related to the aa's distribution}
\usage{
get_dbaasp_data()
}
\description{
This function pulls the data related to the aa's distribution
}
\details{
peptides-complete from \url{https://dbaasp.org/download-dataset?source=peptides}
accessed on 22-04-2022
}
