% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{get_target_additive}
\alias{get_target_additive}
\title{Additive model noise}
\usage{
get_target_additive(kmer_dat, weights = NULL, zero_weight = NULL)
}
\arguments{
\item{kmer_dat}{output of \code{\link{generate_kmer_data}}}

\item{weights}{a vector of weights of motifs' impact on the outcome. The
length of \code{weights} should be the same as the number of motifs provided
during sequences generation (it is the \code{motifs} parameter in the
\code{\link{generate_kmer_data}} function). If \code{weights} oarameter is
\code{NULL}, then weights will be sampled from the uniform distribution on
0-1 interval. The probability of success for target sampling will be
calculated based on the formula provided in details section. Default to
\code{NULL}.}

\item{zero_weight}{a single value denoting the weight of no-motifs case. If
\code{NULL}, then we sample the weight from the uniform distribution on the
[-2, -1] interval. Default to \code{NULL}.}
}
\value{
a binary vector of target variable sampled based on additive model.
}
\description{
This function samples target variable according to the binomial model with
additive impact
}
\details{
This function assumes the following additive binomial model:

\eqn{g(EY) = w_0 + w_1 X_{m_1} + w_2 X_{m_2} + \ldots + w_m X_{m_m}}

where \eqn{w_1, \ldots, w_m} are weights related to motifs.

In the case when \code{weights} is \code{NULL} we calculate the probabilities
based on the formula

\eqn{ exp(1 + x_i)/(1 + exp(1 + x_i))} where xi denotes the sum of weights of
motifs occurring in ith sequence.
}
\examples{
n_seq <- 20
sequence_length <- 20
alph <- letters[1:4]
motifs <- generate_motifs(alph, 4, 4, 4, 6)
results <- generate_kmer_data(n_seq, sequence_length, alph, motifs, n_injections = 4)
get_target_additive(results)

}
