% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{get_target_interactions}
\alias{get_target_interactions}
\title{Interactions model noise}
\usage{
get_target_interactions(kmer_dat, probs = NULL, zero_prob = 0.1)
}
\arguments{
\item{kmer_dat}{output of \code{\link{generate_kmer_data}}}

\item{probs}{an increasing vector of probabilities of success corresponding
to concurrent occurrence of n motifs in a sequence, where n denotes a number
between \code{1} and \code{n_injections}. This vector should have exactly
\code{n_injections} of elements from 0-1 interval. For example, when
\code{n_injections} equals \code{2}, the vector of probabilities should have
two elements, for example, a vector \code{c(0.7, 0.8)} means that we assume
the probability of success equal to \code{0.7} when one motif occurs in a
sequence and \code{0.8} when two motifs occur. The default value is
\code{NULL} meaning that the probabilities will be calculated (See details).}

\item{zero_prob}{a single value denoting the probability of success in the
case when no motifs occur in the sequence. Default to \code{0.1}.}
}
\value{
a binary vector of target variable sampled based on interaction model
and provided/calculated probabilities.
}
\description{
This function samples target variable according to the binomial model with
interactions
}
\details{
This function assumes the following interaction binomial model:

\eqn{g(EY) = w_0 + w_1 (X_{m_1} + X_{m_2} + \ldots + X_{m_k}) +
w_2 \left(\sum_{i = 1}^{k-1}\sum_{j = i + 1}^{k} X_{m_i}X_{m_j}\right) +
\ldots + w_m X_{m_1}\ldots X_{m_m}}

In the case when \code{probs} is \code{NULL} we calculate the probabilities
based on the formula \eqn{ exp(x_i)/(1 + exp(x_i))} where xi denotes the
number of motifs in ith sequence.
}
\examples{
n_seq <- 20
sequence_length <- 20
alph <- letters[1:4]
motifs <- generate_motifs(alph, 4, 4, 4, 6)
results <- generate_kmer_data(n_seq, sequence_length, alph, motifs, n_injections = 4)
get_target_interactions(results)

}
