% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{list_filters}
\alias{list_filters}
\title{Get a list of filtering methods}
\usage{
list_filters()
}
\value{
a character vector of names of filtering functions
}
\description{
Get a list of filtering methods
}
\examples{
list_filters()

}
