% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target.R
\name{rbinom_vec}
\alias{rbinom_vec}
\title{Binomial sampling}
\usage{
rbinom_vec(probs)
}
\arguments{
\item{probs}{a vector of probabilities}
}
\value{
a binary vector of binomial observations corresponding to provided
probabilities.
}
\description{
This function samples from binomial distribution using a vector of
probabilities.
}
\examples{
probs <- runif(100)
rbinom_vec(probs)

}
