% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{sparse_cor}
\alias{sparse_cor}
\title{Sparse correlation}
\usage{
sparse_cor(y_name, xy, candidates_sums, n)
}
\arguments{
\item{y_name}{a character name of reference k-mer}

\item{xy}{a vector of sums of xy products (between y and x(s))}

\item{candidates_sums}{a vector of sums of k-mers}

\item{n}{a numeric number of observations}
}
\value{
a numeric vector of correlations between y and x's.
}
\description{
This function calculates correlation between one k-mer and a matrix of k-mers
under sparsity assumption
}
