% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{validate_motifs}
\alias{validate_motifs}
\title{Validate if given set of motifs can occur in a sequence at the same time}
\usage{
validate_motifs(motifs, sequence_length)
}
\arguments{
\item{motifs}{list of motifs we are checking}

\item{sequence_length}{length of sequence we want to inject}
}
\value{
logical value if such injection is possible
}
\description{
function validates if given motifs can be injected to a sequence of given length
}
\examples{
set.seed(42)
motifs <- generate_motifs(1:4, n_motifs = 2, n_injections = 2, n = 3, d = 3)
validate_motifs(motifs, 7)
validate_motifs(motifs, 9)
}
