% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{count_seq_kmers}
\alias{count_seq_kmers}
\title{Counting kmers}
\usage{
count_seq_kmers(sequences, alphabet, n = 4, d = 6)
}
\arguments{
\item{sequences}{input data for k-mer counting}

\item{alphabet}{elements used to generate a motif}

\item{n}{maximum number of alphabet elements}

\item{d}{number of possible gaps}
}
\description{
This function is a wrapper for seqR counters
}
\examples{
n_seq <- 20
sequence_length <- 10
alph <- letters[1:6]
motifs <- generate_motifs(alph, 3, 3, 3, 2)
seq_data <- generate_sequence_data(n_seq, sequence_length, alph, motifs, 1)
count_seq_kmers(seq_data, alph)

}
