% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_ic}
\alias{filter_ic}
\title{Stepwise information criteria based filtering}
\usage{
filter_ic(target, kmers, ic = "mbic2", reduce = 0.2)
}
\arguments{
\item{target}{a numeric response variable}

\item{kmers}{a matrix of kmers with named columns or an object obtained via
\code{\link{generate_kmer_data}} function.}

\item{ic}{character name of information criterium. One of "aic", "maic",
"bic", "maic2", "mbic", "mbic2". See bigstep package for more information.}

\item{reduce}{a numeric value from (0, 1) interval. Denotes significance
level for preliminary reduction before execution of stepwise procedure.
Default to 0.2.}
}
\value{
a character vector of names of selected kmers
}
\description{
This function is filtering method based on Information Criteria
}
\details{
This function uses bigstep package.
}
\examples{
n_seq <- 200
sequence_length <- 200
alph <- letters[1:4]
motifs <- generate_motifs(alph, 4, 4, 4, 6)
kmers <- generate_kmer_data(n_seq, sequence_length, alph,
                            motifs, n_injections = 4)
target <- get_target_additive(kmers)
filter_ic(target, kmers, "mbic2", 0.2)

}
