% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{generate_sequence}
\alias{generate_sequence}
\title{generates sequence of elements from alphabet with replacement}
\usage{
generate_sequence(sequence_length, alphabet, seqProbs = NULL)
}
\arguments{
\item{sequence_length}{length of generated sample sequence}

\item{alphabet}{elements used to build sequence}

\item{seqProbs}{alphabet probabilites for sequences}
}
\value{
randomly generated sequence
}
\description{
generates sequence of elements from alphabet with replacement
}
\examples{
generate_sequence(5, 1L:4)
generate_sequence(10, c("a", "b", "c"))
generate_sequence(10, c("a", "b", "c"), c(0.6, 0.2, 0.2))
}
