% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{generate_sequence_data}
\alias{generate_sequence_data}
\title{function generates sequences (both positive & negative)}
\usage{
generate_sequence_data(
  n_seq,
  sequence_length,
  alphabet,
  motifs,
  n_injections,
  fraction = 0.5,
  seqProbs = NULL
)
}
\arguments{
\item{n_seq}{number of sequences to be generated}

\item{sequence_length}{sequence length}

\item{alphabet}{elements used to build sequence}

\item{motifs}{list of motifs}

\item{n_injections}{number of motifs injected to each positive sequence}

\item{fraction}{of positive sequences}

\item{seqProbs}{alphabet probabilites for sequences}
}
\value{
generated sequences
}
\description{
function generates sequences (both positive & negative)
}
\examples{
n_seq <- 20
sequence_length <- 10
alph <- 1L:4
motifs <- generate_motifs(alph, 3, 3, 3, 2)
generate_sequence_data(n_seq, sequence_length, alph, motifs, 1)
}
