% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{generate_sequence_data}
\alias{generate_sequence_data}
\title{Sequences generation}
\usage{
generate_sequence_data(
  n_seq,
  sequence_length,
  alphabet,
  motifs,
  n_injections,
  fraction = 0.5,
  seqProbs = NULL
)
}
\arguments{
\item{n_seq}{number of sequences to be generated}

\item{sequence_length}{sequence length}

\item{alphabet}{elements used to build sequence}

\item{motifs}{list of motifs}

\item{n_injections}{maximal number of motifs injected to each positive
sequence (from 1 to \code{n_injections} will be injected)}

\item{fraction}{fraction of positive sequences}

\item{seqProbs}{alphabet probabilities for sequences}
}
\value{
generated sequences with motifs injected
}
\description{
This function generates sequences (both positive & negative).
}
\examples{
n_seq <- 20
sequence_length <- 10
alph <- 1L:4
motifs <- generate_motifs(alph, 3, 3, 3, 2)
generate_sequence_data(n_seq, sequence_length, alph, motifs, 3)

}
